third_party/blink/web_tests/http/tests/devtools/console/console-preserve-scroll-expected.txt

Tests that console preserves scroll position when switching away.

Message count: 100
Console scrollTop: 10

Running: testSwitchToAnotherPanel
Panel sources was opened.
Panel console was opened.
Console scrollTop: 10

Running: testClickLinkToRevealAnotherPanel
Panel sources was opened.
Panel console was opened.
Console scrollTop: 10

Running: testConsolePanelToDrawer
Drawer panel set to console-view
Panel sources was opened.
Console scrollTop: 10
Panel console was opened.
Console scrollTop: 10

Running: testCloseDrawerFromConsolePanelAndOpenFromAnotherPanel
Panel sources was opened.
Drawer panel set to console-view
Console scrollTop: 10
